SEACAS: Sync with seacas repository:
* IOSS: remove surfaceSplitBackwardCompatibility; klockwork cleanup; memory leak
* NEM_SPREAD: Fix memory allocation function to work with large allocations
* NEM_SLICE: Fix memory allocation request to not overflow integer
* NEM_SLICE: Speed up/simplify set intersection routine
* NEM_SPREAD: Remove unused extern function reference
* EPU: Add -sum_shared_nodes option which sums the nodal values of all nodes on processor boundaries (shared nodes)
    This is used for the following case:
    ...discontinuous finite-elements data, which we'd like to combine into continuous nodal data by
       weighting the element data by the mass matrix and summing over
* EJOIN: Update element block construction with new signature (no attribute count)
* NEM_SPREAD,NEM_JOIN: Remove int_cmp; not used and conflicts with netcdf-4
* EXODIFF: More fixes to support 0 as valid entity id
* CONJOIN: Omit outputting time/part information to info records
* NEM_SLICE/NEM_SPREAD: Portability changes
* EXODIFF: Permit 0 as valid set/block/map id; fix compile warnings

+++ packages/seacas/applications/nem_spread/rf_allo.c
@@ -138,17 +138,17 @@ va_dcl
    char *yo = "array_alloc";
    int	i, j;
    struct dim {
-      long	index;	/* Number of elements in the dimension	*/
-      long	total;	/* Total number of elements 		*/
-      long	size;	/* Size of a single element in bytes	*/
-      long	off;	/* offset from beginning of array	*/
+     size_t	index;	/* Number of elements in the dimension	*/
+     size_t	total;	/* Total number of elements 		*/
+     size_t	size;	/* Size of a single element in bytes	*/
+     size_t	off;	/* offset from beginning of array	*/
    }	dim[3];		/* Info about each dimension 		*/
 #ifndef __STDC__
    char *file;		/* Filename of source code from call.   */
    int lineno;		/* Line number of call.                 */
    int numdim;		/* Number of dimensions			*/
 #endif
-   long total;		/* Total size of the array		*/
+   size_t total;		/* Total size of the array		*/
    double *dfield;	/* ptr to avoid lint complaints		*/
    char *field;		/* The multi-dimensional array		*/
    char **ptr;		/* Pointer offset			*/
@@ -210,7 +210,7 @@ va_dcl
 
    total = dim[numdim-1].off + dim[numdim-1].total * dim[numdim-1].size;
 
-   dfield = (double *) smalloc((int) total, file, lineno);
+   dfield = (double *) smalloc(total, file, lineno);
    field = (char *) dfield;
 
    for (i=0; i<numdim - 1; i++) {
